Move function for valid import
